Alterado tipo de int para float no Desafio08

+++ Desafios/Desafio08.py
@@ -3,7 +3,7 @@ Escreva um programa que leia um valor
 em metros e o exiba convertido em centimetros
 e em milimetros
 """
-n = int(input('Digite um valor em metros: '))
+n = float(input('Digite um valor em metros: '))
 cent = n * 100
 mili = n * 1000
 print('O valor de {} metros, corresponde a {} centimetros e em {} milimetros'.format(n, cent, mili))
